feat: mount AI assistant globally in app routes
